refactor 24.10.01 폴더명 변경
